+++ src/main/factory/controller/account/signup/signup-controller-factory.ts
@@ -2,7 +2,7 @@ import { SignUpController } from "../../../../../presentation/controllers/accoun
 import { Controller } from "../../../../../presentation/protocols";
 import { makeSignUpValidation } from "./signup-validation-factory";
 import { makeDbAuthentication } from "../../../usecase/db-authentication-factory";
-import { makeDbAddAccount } from "../../../usecase/db-add-account-factory";
+import { makeDbAddAccount } from "../../../usecase/account/db-add-account-factory";
 
 export const makeSignUpController = (): Controller => {
