+++ src/main/factory/controller/survey/create-survey/create-survey-controller-factory.ts
@@ -1,5 +1,5 @@
 import { CreateSurveyController } from "../../../../../presentation/controllers/survey/create-survey/create-survey-controller";
-import { makeDBAddSurvey } from "../../../usecase/db-add-survey-factory";
+import { makeDBAddSurvey } from "../../../usecase/survey/db-add-survey-factory";
 import { makeCreateSurveyValidation } from "./create-survey-validation-factory";
 
 export const makeCreateSurveyController = (): CreateSurveyController => {
