+++ src/main/factory/usecase/account/db-add-account-factory.ts
@@ -1,0 +1,10 @@
+import { DbAddAccount } from "../../../../data/usecases/add-account/db-add-account";
+import { BcryptAdapter } from "../../../../infra/criptography/bcrypt/bcrypt-adapter";
+import { AccountMongoRepository } from "../../../../infra/db/mongodb/account/account-mongo-repository";
+
+export const makeDbAddAccount = (): DbAddAccount => {
+    const salt = 12;
+    const mongoRepository = new AccountMongoRepository();
+    const dbAddAccount = new DbAddAccount(new BcryptAdapter(salt), mongoRepository, mongoRepository);  
+    return dbAddAccount;
+}
