+++ src/main/factory/usecase/survey/db-add-survey-factory.ts
@@ -1,0 +1,8 @@
+import { DbAddSurvey } from "../../../../data/usecases/add-survey/db-add-survey";
+import { AddSurvey } from "../../../../domain/usecases/add-survey";
+import { SurveyMongoRepository } from "../../../../infra/db/mongodb/survey/survey-mongo-repository";
+
+export const makeDBAddSurvey = (): AddSurvey => {
+    const surveyRepository = new SurveyMongoRepository()
+    return new DbAddSurvey(surveyRepository);
+}
